refactor(decorator): rename field decorator to be consitent with the rename collection
