+++ packages/agent/src/builder/decorators-stack.ts
@@ -9,7 +9,7 @@ import {
   OperatorsReplaceCollectionDecorator,
   PublicationCollectionDecorator,
   RelationCollectionDecorator,
-  RenameCollectionDecorator,
+  RenameFieldCollectionDecorator,
   SchemaCollectionDecorator,
   SearchCollectionDecorator,
   SegmentCollectionDecorator,
@@ -29,7 +29,7 @@ export default class DecoratorsStack {
   lateOpEmulate: DataSourceDecorator<OperatorsEmulateCollectionDecorator>;
   lateOpReplace: DataSourceDecorator<OperatorsReplaceCollectionDecorator>;
   publication: DataSourceDecorator<PublicationCollectionDecorator>;
-  rename: DataSourceDecorator<RenameCollectionDecorator>;
+  renameField: DataSourceDecorator<RenameFieldCollectionDecorator>;
   schema: DataSourceDecorator<SchemaCollectionDecorator>;
   search: DataSourceDecorator<SearchCollectionDecorator>;
   segment: DataSourceDecorator<SegmentCollectionDecorator>;
@@ -69,7 +69,7 @@ export default class DecoratorsStack {
     // Step 4: Renaming must be either the very first or very last so that naming in customer code
     // is consistent.
     last = this.publication = new DataSourceDecorator(last, PublicationCollectionDecorator);
-    last = this.rename = new DataSourceDecorator(last, RenameCollectionDecorator);
+    last = this.renameField = new DataSourceDecorator(last, RenameFieldCollectionDecorator);
     /* eslint-enable no-multi-assign */
 
     this.dataSource = last;
