+++ packages/datasource-toolkit/src/index.ts
@@ -21,7 +21,7 @@ export { default as RelationCollectionDecorator } from './decorators/relation/co
 export { default as OperatorsEmulateCollectionDecorator } from './decorators/operators-emulate/collection';
 export { default as OperatorsReplaceCollectionDecorator } from './decorators/operators-replace/collection';
 export { default as PublicationCollectionDecorator } from './decorators/publication/collection';
-export { default as RenameCollectionDecorator } from './decorators/rename/collection';
+export { default as RenameFieldCollectionDecorator } from './decorators/rename-field/collection';
 export { default as SearchCollectionDecorator } from './decorators/search/collection';
 export { default as WriteCollectionDecorator } from './decorators/write/collection';
 export { default as SchemaCollectionDecorator } from './decorators/schema/collection';
